requirements.txt
python-telegram-bot==20.3
pandas
